backend/requirements.txt
flask 
python-dotenv 
pandas 
scikit-learn 
matplotlib 
seaborn